Allow with_items to operate on hashes
This is a "request for comment" patch.

This patch extends with_items to operate on hashes in addition to lists
by creating a list of {key: value} hashes for use in iteration.

Alternatively, the code could instead return a list of (key,value)
tuples, but there is something pleasing about using explicit names.